refactor: LottoGenerate 클래스를 LottoMachine의 메서드로 변경
