+++ src/domain/LottoMachine.js
@@ -1,11 +1,20 @@
-const LottoGenerator = require("./LottoGenerator");
+const { Random } = require("@woowacourse/mission-utils");
 const { LOTTO_SPEC, ERROR_MESSAGES } = require("../constants");
 
 class LottoMachine {
   buy(money) {
     this.validate(money);
     const amount = Math.floor(money / LOTTO_SPEC.MONEY_UNIT);
-    return Array.from({ length: amount }, LottoGenerator.generate);
+    return Array.from({ length: amount }, this.generate);
+  }
+
+  generate() {
+    const randomNumbers = Random.pickUniqueNumbersInRange(
+      LOTTO_SPEC.MIN_NUMBER,
+      LOTTO_SPEC.MAX_NUMBER,
+      LOTTO_SPEC.LENGTH,
+    );
+    return randomNumbers.sort((numA, numB) => numA - numB);
   }
 
   validate(money) {
